Revert "refactor(runloop): modularize agent run loop"
